requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
urllib3>=2.0.0
lxml>=4.9.0
PyJWT>=2.8.0